# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CampaignForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscribers:
    get:
      operationId: listSubscribers
      tags: [subscribers]
      summary: List subscribers
      parameters:
        - name: workspaceId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: listId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, unsubscribed, bounced, complained]
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Paginated list of subscribers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriberList"
    post:
      operationId: createSubscriber
      tags: [subscribers]
      summary: Create a subscriber
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscriber"
  /subscribers/{subscriberId}:
    get:
      operationId: getSubscriber
      tags: [subscribers]
      summary: Get subscriber by ID
      parameters:
        - name: subscriberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscriber
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscriber"
    patch:
      operationId: updateSubscriber
      tags: [subscribers]
      summary: Update subscriber
      parameters:
        - name: subscriberId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscriber"
    delete:
      operationId: deleteSubscriber
      tags: [subscribers]
      summary: Delete subscriber
      parameters:
        - name: subscriberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /subscribers/import:
    post:
      operationId: importSubscribers
      tags: [subscribers]
      summary: Import subscribers (CSV/JSON)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriberImport"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"

  # ── Lists ────────────────────────────────────────────────────────────────
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List email templates
      parameters:
        - name: workspaceId
          in: query
          required: false
          schema:
            type: integer
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
  /templates/{templateId}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get template by ID
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    patch:
      operationId: updateTemplate
      tags: [templates]
      summary: Update template
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete template
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Automations ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Subscriber:
      type: object
      required: [id, workspaceId, email, status, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, unsubscribed, bounced, complained]
        tags:
          type: array
          items:
            type: string
        attributes:
          type: object
          additionalProperties: true
        source:
          type: ["string", "null"]
        listIds:
          type: array
          items:
            type: integer
        createdAt:
          type: string
  
        updatedAt:
          type: string
    SubscriberList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Subscriber"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    SubscriberInput:
      type: object
      required: [workspaceId, email]
      properties:
        workspaceId:
          type: integer
        email:
          type: string

        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        tags:
          type: array
          items:
            type: string
        listIds:
          type: array
          items:
            type: integer
        attributes:
          type: object
          additionalProperties: true
    SubscriberUpdate:
      type: object
      properties:
        email:
          type: string

        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [active, unsubscribed]
        tags:
          type: array
          items:
            type: string
        attributes:
          type: object
          additionalProperties: true
    SubscriberImport:
      type: object
      required: [workspaceId, subscribers]
      properties:
        workspaceId:
          type: integer
        subscribers:
          type: array
          items:
            $ref: "#/components/schemas/SubscriberInput"
        listIds:
          type: array
          items:
            type: integer
        updateExisting:
          type: boolean
          default: true
    ImportResult:
      type: object
      required: [imported, updated, skipped, errors]
      properties:
        imported:
          type: integer
        updated:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string

    # ── Lists ──────────────────────────────────────────────────────────────
...
    Template:
      type: object
      required: [id, workspaceId, name, category, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: ["integer", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [business, ecommerce, newsletter, event, welcome, transactional, minimal, custom]
        html:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
        isGlobal:
          type: boolean
        createdAt:
          type: string
    TemplateInput:
      type: object
      required: [workspaceId, name, category]
      properties:
        workspaceId:
          type: integer
        name:
          type: string
          minLength: 1
        description:
          type: string
        category:
          type: string
          enum: [business, ecommerce, newsletter, event, welcome, transactional, minimal, custom]
        html:
          type: string
    TemplateUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        category:
          type: string
        html:
          type: string

    # ── Automations ────────────────────────────────────────────────────────
...