# lib/api-zod/src/generated/types/subscriberImport.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CampaignForge API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriberInput } from './subscriberInput';

export interface SubscriberImport {
  workspaceId: number;
  subscribers: SubscriberInput[];
  listIds?: number[];
  updateExisting?: boolean;
}